+++ spec/report_spec.rb
@@ -28,6 +28,11 @@ describe Compendium::Report do
     its(:metrics) { should_not equal report2.metrics }
   end
 
+  describe ".name" do
+    subject { TestReport = Class.new(described_class) }
+    its(:name) { should == :test }
+  end
+
   describe "#run" do
     context do
       let(:report_class) do